[schema] Move schemas to own tool file
