+++ src/tools/MapboxApiBasedTool.schema.ts
@@ -1,0 +1,21 @@
+// Copyright (c) Mapbox, Inc.
+// Licensed under the MIT License.
+
+import { z } from 'zod';
+
+export const OutputSchema = z.object({
+  content: z.array(
+    z.union([
+      z.object({
+        type: z.literal('text'),
+        text: z.string()
+      }),
+      z.object({
+        type: z.literal('image'),
+        data: z.string(),
+        mimeType: z.string()
+      })
+    ])
+  ),
+  isError: z.boolean().default(false)
+});
