+++ src/tools/version-tool/VersionTool.schema.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Mapbox, Inc.
+// Licensed under the MIT License.
+
+import { z } from 'zod';
+
+export const VersionSchema = z.object({});
