+++ src/utils/versionUtils-cjs.cts
@@ -1,3 +1,6 @@
+// Copyright (c) Mapbox, Inc.
+// Licensed under the MIT License.
+
 import { readFileSync } from 'node:fs';
 import path from 'node:path';
