+++ test/tools/MapboxApiBasedTool.test.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Mapbox, Inc.
+// Licensed under the MIT License.
+
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { z } from 'zod';
 import { MapboxApiBasedTool } from '../../src/tools/MapboxApiBasedTool.js';
