+++ test/tools/annotations.test.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Mapbox, Inc.
+// Licensed under the MIT License.
+
 import { describe, it, expect } from 'vitest';
 import { getAllTools } from '../../src/tools/toolRegistry.js';
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
