+++ test/tools/version-tool/VersionTool.test.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Mapbox, Inc.
+// Licensed under the MIT License.
+
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import type { MockedFunction } from 'vitest';
 import { getVersionInfo } from '../../../src/utils/versionUtils.js';
